feat(EMS-870): account sign in - update sign in tests to use a verified account
