+++ e2e-tests/cypress/e2e/journeys/insurance/account/create/create-account-and-verify-email/create-account-and-verify-email.spec.js
@@ -1,7 +1,6 @@
 import { signInPage } from '../../../../../pages/insurance/account/sign-in';
 import { PAGES } from '../../../../../../../content-strings';
 import { INSURANCE_ROUTES as ROUTES } from '../../../../../../../constants/routes/insurance';
-import api from '../../../../../../support/api';
 
 const CONTENT_STRINGS = PAGES.INSURANCE.ACCOUNT.SIGN_IN.ROOT;
 
@@ -28,28 +27,8 @@ context('Insurance - Account - Create - I want the system to generate account ve
   });
 
   describe(`when navigating to the ${VERIFY_EMAIL} URL with a valid token query parameter`, () => {
-    let exporter;
-
-    before(() => {
-      const exporterEmail = Cypress.env('GOV_NOTIFY_EMAIL_RECIPIENT');
-
-      /**
-       * Get the exporter (with verification hash) directly from the API,
-       * so that we can mimic "clicking email verification link" and proceed with the user flow.
-       */
-      api.getExporterByEmail(exporterEmail).then((response) => {
-        exporter = response;
-      });
-    });
-
     it(`should validate the exporter, redirect to ${SIGN_IN.ROOT} and display a success banner`, () => {
-      // mimic "clicking email verification link"
-      cy.navigateToUrl(`${Cypress.config('baseUrl')}${VERIFY_EMAIL}?token=${exporter.verificationHash}`);
-
-      // User should be verified and therefore redirected to the "sign in" page.
-      const expected = `${Cypress.config('baseUrl')}${SIGN_IN.ROOT}`;
-
-      cy.url().should('eq', expected);
+      cy.verifyAccountEmail();
 
       const { successBanner } = signInPage;
