+++ e2e-tests/cypress/support/api/index.js
@@ -29,11 +29,7 @@ const getExporterByEmail = async (email) => {
       throw new Error('Getting exporter by email', { response });
     }
 
-    const { data } = response.body;
-
-    const exporter = data.exporters[0];
-
-    return exporter;
+    return response;
   } catch (err) {
     console.error(err);
